Divide Getter/Setter tests into smaller tests for each type. Add new testcase for type tests.
